Validation commands: First human refactoring
- For many commands, perform a deeper analysis of input data if -debug is specified.
- Modify many data validity checks based on revised rules and nature of desired user terminal messages.

+++ cpp/core/surface/polygon.cpp
@@ -16,10 +16,12 @@
 
 #include "surface/polygon.h"
 
+#include "surface/types.h"
+
 namespace MR::Surface {
 
 template <> bool Polygon<3>::shares_edge(const Polygon<3> &that) const {
-  uint32_t shared_vertices = 0;
+  size_t shared_vertices = 0;
   if (indices[0] == that[0] || indices[0] == that[1] || indices[0] == that[2])
     ++shared_vertices;
   if (indices[1] == that[0] || indices[1] == that[1] || indices[1] == that[2])
